+++ src/attrs.mli
@@ -22,6 +22,7 @@ val jsonschema_ld_description : (Ppxlib.label_declaration, string Location.loc) 
 val jsonschema_td_description : (Ppxlib.type_declaration, string Location.loc) Ppxlib.Attribute.t
 val jsonschema_cd_description : (Ppxlib.constructor_declaration, string Location.loc) Ppxlib.Attribute.t
 val jsonschema_ct_description : (Ppxlib.core_type, string Location.loc) Ppxlib.Attribute.t
+val jsonschema_rtag_description : (Ppxlib.row_field, string Location.loc) Ppxlib.Attribute.t
 val jsonschema_td_format : (Ppxlib.type_declaration, string Location.loc) Ppxlib.Attribute.t
 val jsonschema_ld_format : (Ppxlib.label_declaration, string Location.loc) Ppxlib.Attribute.t
 val jsonschema_ct_format : (Ppxlib.core_type, string Location.loc) Ppxlib.Attribute.t
@@ -35,15 +36,16 @@ val jsonschema_ct_attrs : (Ppxlib.core_type, Ppxlib.expression) Ppxlib.Attribute
 val jsonschema_td_attrs : (Ppxlib.type_declaration, Ppxlib.expression) Ppxlib.Attribute.t
 val jsonschema_ld_attrs : (Ppxlib.label_declaration, Ppxlib.expression) Ppxlib.Attribute.t
 
-(** [ld_description], [td_description], [cd_description] and [ct_description] resolve a
-    description from [\[@jsonschema.description "..."\]]. When [ocaml_doc] is [true] and
-    the explicit annotation is absent, they fall back to an [ocaml.doc] attribute
-    (i.e. a [(** ... *)] comment) on the same node. *)
+(** [ld_description], [td_description], [cd_description], [ct_description] and
+    [rtag_description] resolve a description from [\[@jsonschema.description "..."\]].
+    When [ocaml_doc] is [true] and the explicit annotation is absent, they fall back
+    to an [ocaml.doc] attribute (i.e. a [(** ... *)] comment) on the same node. *)
 val ld_description : ocaml_doc:bool -> Ppxlib.label_declaration -> string Location.loc option
 
 val td_description : ocaml_doc:bool -> Ppxlib.type_declaration -> string Location.loc option
 val cd_description : ocaml_doc:bool -> Ppxlib.constructor_declaration -> string Location.loc option
 val ct_description : ocaml_doc:bool -> Ppxlib.core_type -> string Location.loc option
+val rtag_description : ocaml_doc:bool -> Ppxlib.row_field -> string Location.loc option
 
 val attributes : Ppxlib.Attribute.packed list
 val args : unit -> (bool -> bool -> bool -> 'a, 'a) Ppxlib.Deriving.Args.t